hopefully fixing integer overflow
(unsigned long) should be able to hold (ushort*ushort*3)

CID:27876

+++ plugins/modelOBJ/model_loader.cpp
@@ -2230,7 +2230,7 @@ glmReadPPM(const char* filename, int* width, int* height)
   FILE* fp;
   int i, w, h, d;
   unsigned char* image;
-  size_t imagesize=0;
+  unsigned long imagesize=0;
   char head[70];          /* max line <= 70 in PPM (per spec). */
 
   fp = fopen(filename, "rb");
